setup custom webpack aliases, housekeeping and cleanup
