+++ jsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@utils/*": ["./src/utils/*"],
+			"@components/*": ["./src/components/*"],
+    }
+  },
+	"exclude": ["node_modules", "dist"]
+}
