+++ package.json
@@ -42,6 +42,7 @@
 		"@testing-library/jest-dom": "^6.1.4",
 		"@testing-library/react": "^14.0.0",
 		"@wordpress/scripts": "^26.15.0",
+		"eslint-import-resolver-custom-alias": "^1.3.2",
 		"eslint-webpack-plugin": "^4.1.0",
 		"grunt": "^1.5.3",
 		"grunt-contrib-clean": "^2.0.1",
