+++ src/blocks/looper/components/BlockControls.js
@@ -1,6 +1,6 @@
 import { BlockControls } from '@wordpress/block-editor';
 import { ToolbarButton, ToolbarGroup } from '@wordpress/components';
-import getIcon from '../../../utils/get-icon';
+import getIcon from '@utils/get-icon';
 import { __ } from '@wordpress/i18n';
 import { createBlocksFromInnerBlocksTemplate } from '@wordpress/blocks';
 import { useDispatch } from '@wordpress/data';
