+++ src/blocks/looper/components/LoopInnerBlocksRenderer.js
@@ -1,8 +1,37 @@
-import { useMemo } from '@wordpress/element';
+import {
+	BlockContextProvider,
+	useInnerBlocksProps,
+	__experimentalUseBlockPreview as useBlockPreview, // eslint-disable-line @wordpress/no-unsafe-wp-apis
+	store as blockEditorStore,
+} from '@wordpress/block-editor';
+import { __ } from '@wordpress/i18n';
 import { useSelect } from '@wordpress/data';
-import LoopRenderer from './LoopRenderer';
-import { normalizeRepeatableArgs, removeEmpty } from './utils';
+import { Spinner } from '@wordpress/components';
 import { store as coreStore } from '@wordpress/core-data';
+import { memo, useEffect, useMemo, useState } from '@wordpress/element';
+
+import { useDebouncedCallback } from 'use-debounce';
+
+import { normalizeRepeatableArgs, removeEmpty } from '@utils/index';
+
+function BlockPreview( { blocks } ) {
+	const blockPreviewProps = useBlockPreview( {
+		blocks,
+	} );
+
+	return blockPreviewProps.children;
+}
+
+const MemoizedBlockPreview = memo( BlockPreview );
+
+function setIsBlockPreview( innerBlocks ) {
+	return innerBlocks.map( ( block ) => {
+		const newInnerBlocks = setIsBlockPreview( block.innerBlocks );
+		const attributes = Object.assign( {}, block.attributes, { isBlockPreview: true } );
+
+		return Object.assign( {}, block, { attributes, innerBlocks: newInnerBlocks } );
+	} );
+}
 
 export default function LoopInnerBlocksRenderer( props ) {
 	const {
@@ -45,18 +74,75 @@ export default function LoopInnerBlocksRenderer( props ) {
 		};
 	}, [ JSON.stringify( normalizedQuery ) ] );
 
-	return (
-		<LoopRenderer
-			{ ...props }
-			data={ data }
-			hasData={ !! ( hasResolvedData && data?.length ) }
-			isResolvingData={ isResolvingData }
-			hasResolvedData={ hasResolvedData }
-			clientId={ clientId }
-			contextCallback={ ( post ) => ( {
-				postType: post.type,
-				postId: post.id,
-			} ) }
-		/>
-	);
+	const contextCallback = ( post ) => ( {
+		postType: post.type,
+		postId: post.id,
+	} );
+	const hasData = useMemo( () => !! ( hasResolvedData && data?.length, [ hasResolvedData, data ] ) );
+	const innerBlocks = useSelect( ( select ) => {
+		return select( 'core/block-editor' )?.getBlocks( clientId );
+	}, [] );
+
+	const { getSelectedBlock } = useSelect( blockEditorStore );
+	const [ innerBlockData, setInnerBlockData ] = useState( [] );
+
+	useEffect( () => {
+		setInnerBlockData( setIsBlockPreview( innerBlocks ) );
+	}, [] );
+
+	const debounced = useDebouncedCallback( () => {
+		setInnerBlockData( setIsBlockPreview( innerBlocks ) );
+	}, 10 );
+
+	const debounceBlocks = [
+		'core/paragraph',
+		'core/heading',
+		'core/button',
+		'generateblocks/headline',
+		'generateblocks/button',
+	];
+
+	const innerBlocksProps = useInnerBlocksProps( {}, { templateLock: 'all', renderAppender: () => null }, innerBlockData );
+
+	useEffect( () => {
+		const selectedBlock = getSelectedBlock();
+
+		if (
+			debounceBlocks.includes( selectedBlock?.name ) &&
+			! selectedBlock?.attributes?.useDynamicData &&
+			! selectedBlock?.attributes?.dynamicContentType
+		) {
+			// Only debounce if we're using a RichText component.
+			debounced();
+		} else {
+			setInnerBlockData( setIsBlockPreview( innerBlocks ) );
+		}
+	}, [ JSON.stringify( innerBlocks ) ] );
+
+	const dataContexts = useMemo( () => {
+		if ( hasData && Array.isArray( data ) ) {
+			return data.map( ( item ) => ( contextCallback( item ) ) );
+		}
+
+		return [];
+	}, [ data, hasData ] );
+
+	if ( isResolvingData ) {
+		return ( <Spinner /> );
+	}
+
+	if ( hasResolvedData && ! hasData ) {
+		return ( <h5>{ __( 'No results found.', 'generateblocks' ) }</h5> );
+	}
+
+	return dataContexts && dataContexts.map( ( postContext ) => (
+		<BlockContextProvider key={ postContext.postId } value={ postContext }>
+			{ postContext.postId === dataContexts[ 0 ]?.postId
+				? innerBlocksProps.children
+				: (
+					<MemoizedBlockPreview blocks={ innerBlockData } />
+				)
+			}
+		</BlockContextProvider>
+	) );
 }
