+++ src/utils/index.js
@@ -1,0 +1,15 @@
+import {
+	removeEmpty,
+	normalizeTaxQuery,
+	getTaxQueryParam,
+	normalizeRepeatableArgs,
+	normalizeArgs,
+} from './loop-utils';
+
+export {
+	removeEmpty,
+	normalizeTaxQuery,
+	getTaxQueryParam,
+	normalizeRepeatableArgs,
+	normalizeArgs,
+};
